Fix many nullable warnings in MainWindowPM.cs
The pattern where the constructor calls an Initialize() method is one
that the C# compiler can't analyze to detect that nullable reference
types are actually assigned a real value. So we move most of the code
into the constructor (and the static fields are initializes in a static
constructor) so that the compiler will stop warning us about them.

This still leaves a few compiler warnings that are real, such as the one
about _realDictionaryPath, which could indeed be accessed before it is
initialized. That one, and a few others, should truly be typed nullable.