Get rid of some more nullable warnings
Not all nullable warnings will be able to be removed as long as this
code still targets net461; for example, string.IsNullOrEmpty guarantees
that the string isn't null, but the annotations required for the
compiler to know that weren't added until the .NET Core era. So until
the projects targets a modern version of .NET, the only way to get rid
of nullable references like these is to use the ! operator every time,
which will quickly get tedious.