Added legacy Keyman lib's to the repo as Reference's
This allows the exe to launch. It doesn't appear that legacy-style Reference's are included in deps.json, so when Solid launches with its deps.json it can't find the Keyman dll's, even though they're present in the working directory. Removing deps.json allows Solid to theoretically just look in its working directory, but something in icu.net requires Solid.deps.json.

This solution is pretty unfortunate, but it does work. I'll keep brainstorming for alternatives. One alternative would be to make the three Keyman lib's Nuget packages, but this wasn't done when these were touched a few years ago (for some reason).